requirements.txt
tqdm==4.50.2
torch==2.0.0
PyYAML==6.0
torcheval==0.0.6
hydra-core==1.1.0
safetensors==0.3.2
torchvision==0.15.1
transformers==4.29.1